+++ tests/testthat.R
@@ -1,4 +1,4 @@
-library(testthat)
-library(states)
+library("testthat")
+library("states")
 
 test_check("states")